+++ Cmdlets/Public/Get-TeamViewerLicense.ps1
@@ -1,0 +1,18 @@
+function Get-TeamViewerLicense {
+    param(
+        [Parameter(Mandatory = $true)]
+        [securestring]
+        $ApiToken
+    )
+
+    $resourceUri = "$(Get-TeamViewerApiUri)/company/license"
+
+    $response = Invoke-TeamViewerRestMethod `
+        -ApiToken $ApiToken `
+        -Uri $resourceUri `
+        -Method Get `
+        -WriteErrorTo $PSCmdlet `
+        -ErrorAction Stop
+
+    Write-Output ($response | ConvertTo-TeamViewerLicense)
+}