+++ Cmdlets/Public/Remove-TeamViewerCompany.ps1
@@ -1,0 +1,19 @@
+function Remove-TeamViewerCompany {
+    [CmdletBinding(SupportsShouldProcess = $true)]
+    param(
+        [Parameter(Mandatory = $true)]
+        [securestring]
+        $ApiToken
+    )
+
+    $resourceUri = "$(Get-TeamViewerApiUri)/company"
+
+    if ($PSCmdlet.ShouldProcess('TeamViewer company', 'Delete company')) {
+        Invoke-TeamViewerRestMethod `
+            -ApiToken $ApiToken `
+            -Uri $resourceUri `
+            -Method Delete `
+            -WriteErrorTo $PSCmdlet | `
+            Out-Null
+    }
+}