+++ CHANGELOG.md
@@ -4,7 +4,10 @@
 
 ### Added
 
-- Adds `Get-TeamViewerCompany` that returns the company / tenant details associated with the API token.
+- Adds `Get-TeamViewerCompany` to retrieve the company / tenant details.
+- Adds `Get-TeamViewerLicense` to retrieve company / tenant licenses.
+- Adds `Remove-TeamViewerCompany` to delete the company / tenant (Dangerous!).
+- Adds `Set-TeamViewerCompany` to modify the company / tenant details.
 
 ## 2.6.0 (2026-07-30)